### 2017-09-12 #404 @whedon
[PRE REVIEW]: False arrhythmia alarm reduction in the intensive care unit
**Submitting author:** @liandrea4 (<a href="http://orcid.org/0000-0001-8419-5527">Andrea Li</a>)
**Repository:** <a href="https://github.com/MIT-LCP/false-alarm-reduction" target ="_blank">https://github.com/MIT-LCP/false-alarm-reduction</a>
**Version:** v1.0.1
