### 2017-11-12 #404 @arfon
@stsievert @effigies @oesteban would one of you be willing to review this submission for JOSS?
