### 2017-12-03 #404 @stsievert
I can review this, but probably not right away.
