### 2017-12-04 #404 @arfon
> I can review this, but probably not right away.

Great, thanks @stsievert. No huge rush but I'll go ahead and start the main review issue.
