### 2017-12-04 #481 @whedon
[REVIEW]: False arrhythmia alarm reduction in the intensive care unit
**Submitting author:** @liandrea4 (<a href="http://orcid.org/0000-0001-8419-5527">Andrea Li</a>)
**Repository:** <a href="https://github.com/MIT-LCP/false-alarm-reduction" target ="_blank">https://github.com/MIT-LCP/false-alarm-reduction</a>
**Version:** v1.0.1

- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? [Yes, MIT]
- [x] **Authorship:** Has the submitting author (@liandrea4) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? [Yes, though the commits are listed under Andrew Li]

- [ ] **Installation:** Does installation proceed as outlined in the documentation? https://github.com/MIT-LCP/false-alarm-reduction/pull/2 filed

- [ ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? [No. They have a one sentence description of their code's purpose but no intended audience]
- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. [They mention how to install deps]
- [ ] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). [None seen]
- [ ] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? [I don't see any documentation].
- [ ] **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified? [I don't there are any tests.]
- [ ] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support [Not that I see]

- [ ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? [Not beyond the title/one sentence description]
