### 2017-12-04 #404 @arfon
@stsievert - see you over in #481 for the actual review.
