### 2017-12-22 #481 @stsievert
Sorry for the delay on this, but I've finally got around to starting this review. I've written some comments about, but here I've fleshed some of them out:

* Documentation needs improvement
    * At the very least, I should understand what your software does. Typically I can understand what each function does in a piece of software.
    * I'd have some high-level overview of the software that describes why the software is useful and what features it provides (more than once sentence at the top of the README).
    * Ideally, every functions should have docstrings to describe the inputs/outputs and purpose
    * I'd look at the [readthedocs Getting Started guide][rtd]. Good examples of documentation are with [dask.distributed API] and the [requests API].
* Tests need implementation
    * At minimum, indicate how I can tell if your software performs as expected (likely via a file that uses a couple `assert`ions).
    * Ideally, every function should be tested that it produces the expected outputs. [py.test] is a good tool to help with this.
* Community guidelines
    * I'd have a `CONTRIBUTING` file that lists various tasks you'd like contributors to do. See [GitHub's Contributing Guidelines][gcg].

I'd look at the [JOSS author guidelines] and [JOSS reviewer guidelines].

[requests API]:http://docs.python-requests.org/en/master/
[dask.distributed API]:https://distributed.readthedocs.io/en/latest/
[JOSS author guidelines]:http://joss.theoj.org/about#author_guidelines
[JOSS reviewer guidelines]:http://joss.theoj.org/about#reviewer_guidelines
[gcg]:https://github.com/blog/1184-contributing-guidelines
[rtd]:https://docs.readthedocs.io/en/latest/getting_started.html
[py.test]:https://docs.pytest.org/en/latest/
