+++ pyfar/classifier.py
@@ -1,10 +1,12 @@
+from __future__ import print_function
+
 from spectrum                   import *
 from scipy.stats                import kurtosis
 from sklearn.linear_model       import LogisticRegression, LassoCV
 from sklearn.metrics            import auc, roc_curve
 from datetime                   import datetime
 import numpy                    as np
-import matplotlib.pyplot  		as plt
+import matplotlib.pyplot          as plt
 import csv
 import wfdb
 
@@ -32,7 +34,7 @@ def get_psd(channel_subsig, order, nfft):
     # plt.xlabel('Normalized Frequency')
     # plt.show()
 
-    # print len(psd)
+    # print(len(psd))
 
     return psd
 
@@ -106,7 +108,7 @@ def get_channel_type(channel_name):
             channel_type = splitted_line[0]
 
             if channel_name == channel:
-            	return channel_type
+                return channel_type
 
     raise Exception("Unknown channel name")
 
@@ -149,8 +151,8 @@ def generate_features(features_filename):
                 ecg_channels = get_channels_of_type(fields['signame'], "ECG")
 
                 if len(ecg_channels) == 0:
-    		    	print "NO ECG CHANNELS FOR SAMPLE: ", sample_name
-    		    	continue
+                    print("NO ECG CHANNELS FOR SAMPLE: {}".format(sample_name))
+                    continue
 
                 channel_subsig = subsig[:,int(ecg_channels[0])]
 
@@ -161,11 +163,11 @@ def generate_features(features_filename):
                     pursqi = get_pursqi(channel_subsig)
 
                 except Exception as e:
-                    print "sample_name:", sample_name, e
+                    print("sample_name: {}\n{}".format(sample_name, e))
                     continue
 
                 if np.isnan([baseline, power, ksqi, pursqi]).any():
-                    print "sample containing nan:", sample_name, [baseline, power, ksqi, pursqi]
+                    print("sample containing nan: {}\n{}".format(sample_name, [baseline, power, ksqi, pursqi]))
                     continue
 
                 if int(sample_number) < TRAINING_THRESHOLD:
@@ -223,7 +225,7 @@ def get_score(prediction, true):
     FP = np.sum([(prediction[i] == 1) and (true[i] == 0) for i in range(len(prediction))])
     FN = np.sum([(prediction[i] == 0) and (true[i] == 1) for i in range(len(prediction))])
 
-    # print TP, TN, FP, FN
+    # print('{} {} {} {}'.format(TP, TN, FP, FN))
 
     numerator =  TP + TN
     denominator = FP + 5*FN + numerator
@@ -231,76 +233,81 @@ def get_score(prediction, true):
     return float(numerator) / denominator
 
 
-# print "Generating datasets..."
-# # generate_features(features_filename)
-# training_x, training_y, testing_x, testing_y = generate_datasets(features_filename)
 
-# print len(training_y), len(testing_y)
+if __name__ == '__main__':
+    print("Nothing to do!")
+    # print("Generating datasets...")
+    # # generate_features(features_filename)
+    # training_x, training_y, testing_x, testing_y = generate_datasets(features_filename)
 
-
-#     # start = datetime.now()
-#     # print "Starting at", start
-#     # print "Generating datasets..."
-#     # training_x, training_y, testing_x, testing_y = generate_training_testing()
+    # print("{} {}".format(len(training_y), len(testing_y)))
 
 
-
[…]
