+++ pyfar/ventricular_beat_stdev.py
@@ -1,3 +1,5 @@
+from __future__ import print_function
+
 from classifier             import get_baseline, get_power, get_ksqi, get_pursqi
 from fastdtw                import fastdtw
 from scipy.spatial.distance import euclidean
@@ -24,11 +26,11 @@ MAX_PEAK_DIST = 60. / 30 * 250
 
 DEBUG = False
 
-def dprint(*args): 
-    if DEBUG: 
-        for arg in args: 
-            print arg,
-        print ""
+def dprint(*args):
+    if DEBUG:
+        for arg in args:
+            print(arg)
+        print("")
 
 
 def is_noisy(
@@ -38,7 +40,7 @@ def is_noisy(
         power_threshold=0.9,
         ksqi_threshold=4,
         pursqi_threshold=5
-    ): 
+    ):
 
     checks = []
     dprint(get_baseline(channel_subsig), get_power(channel_subsig), get_ksqi(channel_subsig))
@@ -49,21 +51,21 @@ def is_noisy(
     ksqi_check = get_ksqi(channel_subsig) > ksqi_threshold
     # pursqi_check = get_pursqi(channel_subsig) > pursqi_threshold
     # checks = [baseline_check, power_check, ksqi_check, pursqi_check]
-    
+
     # TODO: maybe high pass filter instead of using baseline check as a check
-    if 'baseline' in checks_to_use: 
+    if 'baseline' in checks_to_use:
         checks.append(baseline_check)
 
-    if 'power' in checks_to_use: 
+    if 'power' in checks_to_use:
         checks.append(power_check)
 
-    if 'ksqi' in checks_to_use: 
+    if 'ksqi' in checks_to_use:
         checks.append(ksqi_check)
 
     return not all(checks)
 
 
-def get_adjusted_ann_indices(annotation, ann_index, start_ratio=1/3.): 
+def get_adjusted_ann_indices(annotation, ann_index, start_ratio=1/3.):
     a = annotation[ann_index-1]
     b = annotation[ann_index]
     c = annotation[ann_index+1]
@@ -76,58 +78,58 @@ def get_adjusted_ann_indices(annotation, ann_index, start_ratio=1/3.):
     return ann_start_index, ann_end_index
 
 
-## 
-#  Returns self_beats, a list of: 
+##
+#  Returns self_beats, a list of:
 #       annotation index
 #       beat_sig
 #  for regular beats detected in own patient's signal
 ##
 def get_self_beats(
-        channel_sig, 
-        annotation, 
-        sample_name, 
+        channel_sig,
+        annotation,
+        sample_name,
         checks_to_use=['baseline', 'power', 'ksqi'],
         num_self_beats=20,
-        window_increment=10, 
-        fs=250.): 
+        window_increment=10,
+        fs=250.):
 
     self_beats = []
 
     # Get self beats in first 2 min
-    for start_time in range(0, 120-window_increment+1, window_increment): 
+    for start_time in range(0, 120-window_increment+1, window_increment):
         end_time = start_time + window_increment
         start_index = int(start_time * fs)
         end_index = int(end_time * fs)
 
         channel_subsig = channel_sig[start_index:end_index]
-        # print start_index, end_index,
+        # print(start_index, end_index,)
 
-        if not is_noisy(channel_subsig, checks_to_use): 
+        if not is_noisy(channel_subsig, checks_to_use):
             for ann_index in range(1, len(annotation)-1):
                 # TODO: update to have the start and end index be smoothed over past values
                 ann_start_index, ann_end_index = get_adjusted_ann_indices(annotation, ann_index)
 
                 # If beat annotation in clean (not noisy) data range
-                if ann_start_index > start_index and ann_end_index < end_index: 
[…]
