+++ README.rst
@@ -1,0 +1,163 @@
+false-alarm-reduction
+=====================
+
+|DOI|
+
+Code for building a model to reduce false alarms in the intensive care
+unit.
+
+Detailed Installation (Ubuntu 16.04)
+------------------------------------
+
+Install Virtual Environment (Python 2.7)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+We recommend installing the package in a virtual environment. If not
+using a virtual environment, skip ahead to the
+``pip install pip --upgrade`` step.
+
+First, install pip and virtualenv as follows:
+
+::
+
+    sudo apt-get install python-pip python-dev python-virtualenv
+
+Create a virtual environment
+
+::
+
+    virtualenv --system-site-packages TARGET_DIRECTORY
+
+Where TARGET\_DIRECTORY is the desired location of the virtual
+environment. Here, we assume it is \`\ ``~/false-alarm-reduction``.
+
+Activate the virtual environment
+
+::
+
+    source ~/false-alarm-reduction/bin/activate
+
+Now you should be working in the virtual environment. Verify pip is
+installed:
+
+::
+
+    (false-alarm-reduction)$ easy_install -U pip
+
+Upgrade pip.
+
+::
+
+    (false-alarm-reduction)$ pip install pip --upgrade
+
+Now install all the necessary packages using the requirements file (for
+reference, these are: ``wfdb``, ``numpy``, ``scipy``, ``matplotlib``,
+``sklearn``, ``fastdtw``, ``spectrum``, ``peakutils``).
+
+::
+
+    pip install -r requirements.txt
+
+Download data
+~~~~~~~~~~~~~
+
+Two convenience scripts are provided to download the data. Run these
+from the main folder as follows:
+
+::
+
+    sh download/download_annotations.sh
+    sh download/download_data.sh
+
+These scripts download the following data:
+
+-  training.zip https://physionet.org/challenge/2015/training.zip
+-  ann\_gqrs0.zip
+   https://www.dropbox.com/sh/hv4uat0ihwlygq8/AABvdXbSGZi3COPG-O\_-nBGxa/ann\_gqrs0.zip?dl=1
+-  ann\_gqrs1.zip
+   https://www.dropbox.com/sh/hv4uat0ihwlygq8/AAAAb14a\_NN8iKojXEoInXCGa/ann\_gqrs1.zip?dl=1
+-  ann\_wabp.zip
+   https://www.dropbox.com/sh/hv4uat0ihwlygq8/AAALSmteHaL0gQovwXj8CXV4a/ann\_wabp.zip?dl=1
+-  ann\_wpleth.zip
+   https://www.dropbox.com/sh/hv4uat0ihwlygq8/AAAko1RNvgmdhWF7lNux-Ob3a/ann\_wpleth.zip?dl=1
+
+Afterward, you should have the following directory structure:
+
+-  ``annotations`` subfolder with all annotations (*.gqrs0, *.gqrs1,
+   *.wabp, *.wpleth)
+-  ``data/training`` subfolder with all data (*.mat and *.hea) and a
+   RECORDS file
+
+Brief instructions for other OS
+-------------------------------
+
+1. Download and install the following packages: ``wfdb``, ``numpy``,
+   ``scipy``, ``matplotlib``, ``sklearn``, ``fastdtw``, ``spectrum``,
+   ``peakutils``.
+2. Download data and annotations
+
+-  training.zip https://physionet.org/challenge/2015/training.zip
+-  ann\_gqrs0.zip
+   https://www.dropbox.com/sh/hv4uat0ihwlygq8/AABvdXbSGZi3COPG-O\_-nBGxa/ann\_gqrs0.zip?dl=1
+-  ann\_gqrs1.zip
[…]
