+++ setup.py
@@ -1,0 +1,90 @@
+# Always prefer setuptools over distutils
+from setuptools import setup, find_packages
+
+# To use a consistent encoding
+from codecs import open
+from os import path
+
+here = path.abspath(path.dirname(__file__))
+
+# Get the long description from the README file
+with open(path.join(here, 'README.rst'), encoding='utf-8') as f:
+    long_description = f.read()
+
+setup(
+    name='pyfar',
+
+    # Versions should comply with PEP440. For a discussion on single-sourcing
+    # the version across setup.py and the project code, see
+    # https://packaging.python.org/en/latest/single_source_version.html
+    version='0.1.0',
+
+    description='False Alarm Reduction for physiologic waveform alarms',
+    long_description=long_description,
+
+    # The project's main homepage.
+    url='https://github.com/MIT-LCP/false-alarm-reduction',
+
+    # Author details
+    author='Andrea Li',
+    author_email='liandrea@mit.edu',
+
+    # Choose your license
+    license='MIT',
+
+    # What does your project relate to?
+    keywords='false alarm reduction ICU',
+
+    # You can just specify the packages manually here if your project is
+    # simple. Or you can use find_packages().
+    packages=find_packages(exclude=['contrib', 'docs', 'tests']),
+
+    # List run-time dependencies here.  These will be installed by pip when
+    # your project is installed. For an analysis of "install_requires" vs pip's
+    # requirements files see:
+    # https://packaging.python.org/en/latest/requirements.html
+    install_requires=[
+         'fastdtw>=0.3.2'
+        ,'matplotlib>=2.0.2'
+        ,'numpy>=1.13.1'
+        ,'pandas>=0.20.3'
+        ,'PeakUtils>=1.1.0'
+        ,'scikit-learn>=0.19.0'
+        ,'scipy>=0.19.0'
+        ,'sklearn>=0.0'
+        ,'spectrum>=0.7.1'
+        ,'wfdb==1.2.2'
+        ],
+
+
+
+    # List additional groups of dependencies here (e.g. development
+    # dependencies). You can install these using the following syntax,
+    # for example:
+    # $ pip install -e .[dev,test]
+    # extras_require={
+    #     'dev': ['check-manifest'],
+    #     'test': ['coverage'],
+    # },
+
+    # If there are data files included in your packages that need to be
+    # installed, specify them here.  If using Python 2.6 or less, then these
+    # have to be included in MANIFEST.in as well.
+    package_data={'pyfar': ['sigtypes', 'vtach_beats.csv']},
+
+    # Although 'package_data' is the preferred approach, in some case you may
+    # need to place data files outside of your packages. See:
+    # http://docs.python.org/3.4/distutils/setupscript.html#installing-additional-files # noqa
+    # In this case, 'data_file' will be installed into '<sys.prefix>/my_data'
+    # data_files=[('my_data', ['data/data_file'])],
+
+    # To provide executable scripts, use entry points in preference to the
+    # "scripts" keyword. Entry points provide cross-platform support and allow
+    # pip to create the appropriate form of executable for the target platform.
+    # entry_points={
+    #     'console_scripts': [
+    #         'sample=sample:main',
+    #     ],
+    # },
+
+)
