### 2018-04-04 980e97a
remove info that has been moved to docs
