+++ docs/pyfar.rst
@@ -1,0 +1,34 @@
+pyfar
+===================
+
+Classifier
+-------------------
+.. automodule:: baseline_algorithm
+  :members:
+
+.. automodule:: classifier
+  :members:
+
+Utilities
+-------------------
+.. automodule:: utils
+  :members:
+
+DTW
+-------------------
+.. automodule:: dtw
+  :members:
+
+Pipeline
+-------------------
+.. automodule:: pipeline
+  :members:
+
+Ventricular beat modelling
+--------------------------
+
+.. automodule:: ventricular_beat_stdev
+  :members:
+
+.. automodule:: ventricular_beat_bank
+  :members:
