+++ docs/index.rst
@@ -1,0 +1,27 @@
+.. False Alarm Reduction documentation master file, created by
+   sphinx-quickstart on Tue Feb 13 09:57:24 2018.
+
+False Alarm Reduction
+=====================
+
+False Alarm Reduction is a library for reducing the number of false alarms
+when detecting events in physiologic waveforms.
+
+See :doc:`the quickstart <quickstart>` to get started.
+
+Contents
+--------
+
+.. toctree::
+  :maxdepth: 1
+  :caption: Getting Started
+
+  quickstart
+  install
+  contributing
+
+.. toctree::
+  :maxdepth: 2
+  :caption: Package Reference
+
+  pyfar
