### 2018-04-04 #481 @alistairewj
Hey @arfon - thanks for patiently following up. We have done a lot - the docs have been rewritten from scratch more or less and are now much more organized - using Sphinx with the plan to host them on readthedocs. Also tidied up a lot of the functions to be more modular and testable - though no tests so far. It's much improved and many thanks to @stsievert for the review.

However, I'm starting to believe we don't really have the time to fully flesh out the package. Essentially, re-reading the journal guidelines, I think our package is too narrow of a contribution (more or less 1-2 algorithms for false alarm reduction), and I don't think we have the time to expand it outside of that (making it more modular and extensible for other users). @liandrea4 is also very busy with getting her masters in an unrelated project. I'm leaning towards deploying the docs this weekend and not pursuing publication - but would be great to hear your thoughts. Thanks again.
