### 2018-04-10 #481 @arfon
> However, I'm starting to believe we don't really have the time to fully flesh out the package. Essentially, re-reading the journal guidelines, I think our package is too narrow of a contribution (more or less 1-2 algorithms for false alarm reduction), and I don't think we have the time to expand it outside of that (making it more modular and extensible for other users). @liandrea4 is also very busy with getting her masters in an unrelated project. I'm leaning towards deploying the docs this weekend and not pursuing publication - but would be great to hear your thoughts. Thanks again.

@alistairewj - thanks for getting back to us - I totally understand where you're coming from. If you like we can withdraw this submission on your behalf. If you ever decide to revisit this we'd welcome a re-submission.
